add checks for text comment and commentId

+++ src/services/commentsService.ts
@@ -1,12 +1,20 @@
+import { DBError } from "../controllers/DBError";
 import { commentsRepository } from "../repositories/commentsRepository";
 import type {
 	DeleteComment,
 	GetAllComments,
 	SendComment,
 } from "../types/services/commentsService";
+import { HTTP_STATUSES } from "../utils";
 
 export const commentsService = {
 	async sendComment(userId: string, data: SendComment) {
+		if (data.text.trim() === "") {
+			throw new DBError(
+				"Сообщение не должно быть пустым",
+				HTTP_STATUSES.BAD_REQUEST_400,
+			);
+		}
 		return await commentsRepository.sendMessage(userId, data);
 	},
 
@@ -15,6 +23,12 @@ export const commentsService = {
 	},
 
 	async deletedComment(userId: string, data: DeleteComment) {
+		if (data.commentId.trim() === "") {
+			throw new DBError(
+				"ID сообщения не должно быть пустым",
+				HTTP_STATUSES.BAD_REQUEST_400,
+			);
+		}
 		return await commentsRepository.deleteComment(userId, data);
 	},
 };
